Wire pending-selection badge click to start a new thread
Extract the local useStartNewThread hook from HoveringToolbar into a
shared module so the pending-selection icon badge can fire the same
'open new-thread form for the current selection' action when clicked.

HoveringToolbar's comment button keeps using the same hook for now —
removed in a follow-up.
